move config files
