+++ src/view/canvas_components/accessory.ts
@@ -5,7 +5,7 @@ import {Point} from '../../utils/index';
 
 import {CanvasManager} from './canvas_manager';
 import {Joint} from './joint';
-import {RenderOrderConfig} from './render_order_config';
+import {RenderOrderConfig} from '../../config/render_order_config';
 import * as ViewFactory from './view_factory';
 import {AccessoryExportedProperties, ExportedProperties, ObjectOptions, ViewObject} from './view_object';
 import {Wall} from './wall';
