+++ src/view/canvas_components/background.ts
@@ -1,7 +1,7 @@
 import {fabric} from 'fabric';
 
 import {CanvasManager} from './canvas_manager';
-import {RenderOrderConfig} from './render_order_config';
+import {RenderOrderConfig} from '../../config/render_order_config';
 import {ExportedProperties, ObjectOptions, ViewObject} from './view_object';
