+++ src/view/canvas_components/canvas_manager.ts
@@ -3,7 +3,7 @@ import {fabric} from 'fabric';
 import {GetImage} from '../../image_manager';
 import {Point} from '../../utils/index';
 
-import {RenderOrderConfig} from './render_order_config';
+import {RenderOrderConfig} from '../../config/render_order_config';
 import * as ViewFactory from './view_factory';
 import {ViewObject} from './view_object';
