+++ src/view/canvas_components/furniture_group.ts
@@ -4,7 +4,7 @@ import {GetImage} from '../../image_manager';
 import {Point} from '../../utils/index';
 
 import {CanvasManager} from './canvas_manager';
-import {RenderOrderConfig} from './render_order_config';
+import {RenderOrderConfig} from '../../config/render_order_config';
 import * as ViewFactory from './view_factory';
 import {FurnitureExportedProperties, ObjectOptions, ViewObject} from './view_object';
