+++ src/view/canvas_components/room.ts
@@ -2,7 +2,7 @@ import {fabric} from 'fabric';
 
 import {CanvasManager} from './canvas_manager';
 import {Joint} from './joint';
-import {RenderOrderConfig} from './render_order_config';
+import {RenderOrderConfig} from '../../config/render_order_config';
 import * as ViewFactory from './view_factory';
 import {ObjectOptions, PROPERTY_TYPE_OPTION, RoomExportedProperties, ViewObject} from './view_object';
